+++ core/AutoLoadPluginThread.py
@@ -27,6 +27,72 @@ syntaxHighlighter Struct:
 """
 
 
+class OnePluginLoadThread(QThread):
+    loaded = Signal()
+    setName = Signal(str)
+
+    def __init__(self, filePath: str) -> None:
+        super().__init__()
+        self.filePath = filePath
+
+    def run(self) -> None:
+        item: Path = Path(self.filePath)
+        self.setName.emit(item.name)
+        debugPluginLog(f"Loading {item.name}")
+        if not item.is_dir():
+            Logger.info(f"Automatic skipped {item.name}, Reason: not a folder ❌")
+            return
+
+        infoJson: Path | PurePath = item / "info.json"
+        if not (infoJson.exists()):
+            Logger.warning(
+                f"Automatic skipped {item.name}, Reason: info.json not exists ❌"
+            )
+            return
+        temp = LoadPluginInfo(item.name).getValue()
+        debugPluginLog(f"Get value: {temp}")
+        if temp == -1:
+            Logger.error(
+                f"Failed to load plugin that its name is {item.name}. Error occurred."
+            )
+            return
+        elif not isinstance(temp, list):
+            Logger.error(f"Returned value is not a list! Plugin name: {item.name}")
+            return
+        loadedPlugin[temp[0]["objectName"]] = temp[0]
+        self.setName.emit(temp[0]["name"])
+        if temp[0]["objectName"] in settingObject.getValue("disableplugin"):
+            debugPluginLog(f"Automatic skip plugin {temp[0]["objectName"]} (DISABLED)")
+            self.loaded.emit()
+            return
+        debugPluginLog(
+            f"Successfully loaded {item.name}, objectName: {temp[0]["objectName"]}. Preparing to parse... ✅"
+        )
+        for key in temp[1]:
+            debugPluginLog(f"Loading {key[0]}...")
+            if key[1] == 1:
+                debugPluginLog("Checked its property! Type: SyntaxHighlighter 🔎")
+                before: datetime = datetime.now()
+                syntaxHighlighter[key[2]] = [key[4], key[3], key[5], key[6]]
+            elif key[1] == 0:
+                debugPluginLog("Checked its property! Type: RunningFunc 🤓")
+                debugPluginLog("Appending to autoRun... 💥")
+                if isinstance(key[2], list):
+                    [
+                        autoRun.append(i) if isinstance(i, partial) else None
+                        for i in key[2]
+                    ]
+                """
+                This code definitely equals
+                for i in key[2]:
+                    if isinstance(i, partial):
+                        autoRun.append(i)
+                But it's will create a u\nused list ([None if isinstance(i, partial) for i in key[2]])
+                """
+                debugPluginLog("Successfully to append! ✅")
+        self.loaded.emit()
+
+
 class AutoLoadPlugin(QThread):
     loadedOne = Signal()
     loadNameChanged = Signal(str)
@@ -56,63 +122,16 @@ class AutoLoadPlugin(QThread):
         self.loadTotal.emit(len(dirs))
         debugPluginLog(f"Total: {len(dirs)}, Starting load... 💥")
         beforeLoadDatetime: datetime = datetime.now()
+        plugins: list[QThread] = []
         for item in dirs:
-            debugPluginLog(f"Loading {item.name}")
-            if not item.is_dir():
-                Logger.info(f"Automatic skipped {item.name}, Reason: not a folder ❌")
-                continue
-
-            infoJson: Path | PurePath = item / "info.json"
-            if not (infoJson.exists()):
-                Logger.warning(
-                    f"Automatic skipped {item.name}, Reason: info.json not exists ❌"
-                )
-                continue
-            self.loadNameChanged.emit(item.name)
-            temp = LoadPluginInfo(item.name).getValue()
-            debugPluginLog(f"Get value: {temp}")
-            if temp == -1:
-                Logger.error(
-                    f"Failed to load plugin that its name is {item.name}. Error occurred."
-                )
-                continue
-            elif not isinstance(temp, list):
-                Logger.error(f"Returned value is not a list! Plugin name: {item.name}")
-                continue
-            loadedPlugin[temp[0]["objectName"]] = temp[0]
-            self.loadNameChanged.emit(temp[0]["name"])
-            if temp[0]["objectName"] in settingObject.getValue("disableplugin"):
-                debugPluginLog(
-                    f"Automatic skip plugin {temp[0]["objectName"]} (DISABLED)"
-                )
-                self.loadedOne.emit()
-                continue
-            debugPluginLog(
-                f"Successfully loaded {item.name}, objectName: {temp[0]["objectName"]}. Preparing to parse... ✅"
-            )
-            for key in temp[1]:
-                debugPluginLog(f"Loading {key[0]}...")
-                if key[1] == 1:
-                    debugPluginLog("Checked its property! Type: SyntaxHighlighter 🔎")
-                    before: datetime = datetime.now()
-                    syntaxHighlighter[key[2]] = [key[4], key[3], key[5], key[6]]
-                elif key[1] == 0:
-                    debugPluginLog("Checked its property! Type: RunningFunc 🤓")
-                    debugPluginLog("Appending to autoRun... 💥")
-                    if isinstance(key[2], list):
-                        [
-                            autoRun.append(i) if isinstance(i, partial) else None
-                            for i in key[2]
-                        ]
-                    """
-                    This code definitely equals
-                    for i in key[2]:
-                        if isinstance(i, partial):
-                            autoRun.append(i)
-                    But it's will create a unused list ([None if isinstance(i, partial) for i in key[2]])
-                    """
-                    debugPluginLog("Successfully to append! ✅")
-            self.loadedOne.emit()
+            thread: OnePluginLoadThread = OnePluginLoadThread(item)
+            thread.start()
+            thread.loaded.connect(self.loadedOne.emit)
+            thread.setName.connect(self.loadNameChanged.emit)
+            debugPluginLog(f"Released thread for loading plugin {item.name}")
+            plugins.append(thread)
+        for item in plugins:
+            item.wait()
         usedTime: float = (datetime.now() - beforeLoadDatetime).total_seconds()
         debugPluginLog(f"Successfully to load plugins! ✅ Used {usedTime:.3f}s")
         assessment = ""