+++ main.py
@@ -1,6 +1,6 @@
 """
 Sinote: Code Smartly, User Interface Intuitive
-Sinote Version 0.06.26017
+Sinote Version 0.06.26024
 3rd Party License saved at ./3rd-party-license
 
 Win12Home (C) 2025, 2026. MIT License.
@@ -14,7 +14,7 @@ AI help me to finish 15% of these code, WTF (BABABOY)
 
 from core.addons.asciiOutput import asciiOutput
 from core.excepthook import setExceptionHook
-from utils.signal import analyzeAllSignal
+from utils.signalExecute import analyzeAllSignal
 from utils.sinoteScript import appStart
 from utils.verCheck import verCheck