+++ ui/edit/AutomaticSaveThingsThread.py
@@ -36,7 +36,7 @@ class AutomaticSaveThingsThread(QThread):
     def run(self) -> None:
         while self.running:
             for temp in range(self.saveSecs * 10):
-                time.sleep(0.1)
+                self.msleep(100)
                 if not self.running:
                     self.saveThings()
                     break