+++ utils/signalExecute.py
@@ -3,7 +3,7 @@ Signal Analyzer
 """
 
 import os
-import signal
+import signal as sig
 from datetime import datetime
 from typing import Any
 
@@ -19,7 +19,7 @@ def executeInterruptSignal(*args: Any) -> None:  # NOQA, args was unused
     global interruptTime
     if (datetime.now() - interruptTime).total_seconds() >= 5:
         Logger.error(
-            "Don't use Ctrl+C to quit forcibly, else you wanna be forgot your data",
+            "Don't use Ctrl+C to quit forcibly, else you wanna unuse your data",
             "SignalAnalyzer",
         )
         Logger.warning("If you want, please try Ctrl+C again", "SignalAnalyzer")
@@ -29,12 +29,12 @@ def executeInterruptSignal(*args: Any) -> None:  # NOQA, args was unused
         "Quiting Sinote forcibly with SIGINT exit code 130...", "SignalAnalyzer"
     )
     if QApplication.instance():
-        QApplication.instance().quit()
+        QApplication.instance().deleteLater()
     os._exit(130)
 
 
 def analyzeInterruptSignal() -> None:
-    signal.signal(signal.SIGINT, executeInterruptSignal)
+    sig.signal(sig.SIGINT, executeInterruptSignal)
 
 
 def executeTerminateSignal(*args: Any) -> None:  # NOQA, args was unused
@@ -45,11 +45,11 @@ def executeTerminateSignal(*args: Any) -> None:  # NOQA, args was unused
 
 
 def analyzeTerminateSignal() -> None:
-    signal.signal(signal.SIGTERM, executeTerminateSignal)
+    sig.signal(sig.SIGTERM, executeTerminateSignal)
 
 
-def executePipeKillSignal(*args: Any, isSIGHUP: bool) -> None:  # NOQA, args was unused
-    Logger.error("Sinote has been killed by PIPE", "SignalAnalyzer")
+def executePipeKillSignal(isSIGHUP: bool) -> None:  # NOQA, args was unused
+    Logger.error("Sinote has been killed by PIPE/HUP signal", "SignalAnalyzer")
     if QApplication.instance():
         QApplication.instance().deleteLater()
     os._exit(129 if isSIGHUP else 141)
@@ -57,9 +57,9 @@ def executePipeKillSignal(*args: Any, isSIGHUP: bool) -> None:  # NOQA, args was
 
 def analyzePipeKillSignal() -> None:
     try:
-        signal.signal(signal.SIGHUP, lambda *args: executePipeKillSignal(args, True))
-        signal.signal(signal.SIGPIPE, lambda *args: executePipeKillSignal(args, False))
-    except:
+        sig.signal(sig.SIGHUP, lambda *args: executePipeKillSignal(isSIGHUP=True))  # NOQA
+        sig.signal(sig.SIGPIPE, lambda *args: executePipeKillSignal(isSIGHUP=False))  # NOQA
+    except (AttributeError, TypeError):
         pass