+++ CHANGELOG.md
@@ -3,10 +3,17 @@
 ## Recently Builds
 ***Now no any build because only Pre-release out!***
 ## Recently Pre-release
-### 0.06.26017
+### 0.06.26024
 1. **Fix issue**: Fixed that Language Setting Widget it won't follow the setting of setting.json5, and baseInfo always be English.
-2. **Add feature**: Added `Add File, Add Folder, Rename, Remove` to the right clicking menu of the tree view of the project.
+2. **Add feature**: Added `Add File, Add Folder, Rename, Remove` to the right-clicking menu of the tree view of the project.
 3. **Add feature**: Added Number Highlighting
+4. **Optimize**: Added ThreadPool to load plugins quickly.
+5. **Fix issue**: Fixed issue that frameless window has not border.
+6. **Fix issue**: Fixed issue that cannot drag window in Wayland and cannot zoom in when the window has been dragging in **GNOME**.
+7. **Fix issue**: Fixed issue that signal import error. (utils.signal is same as signal library, so just rename, ok?)
+8. **Optimize**: Replaced MD5 to XXHASH for speed yeah.
+9. **Fix issue**: Fixed issue that new project window can be maximized with double-clicking.
+10. **Fix issue**: Fixed issue that maximized window won't be restored to maximized at next starting.
 ### 0.06.26014
 1. **Something not important**: Add log output when loading header!
 2. **Code change**: Change function `addLog` to object `Logger` for readability.